Created a protocol file, CalcEngine conforms to it.
* Not all of the public functions in CalcEngine are loaded into the
protocol. Look for this in the next commit.
